Clean unused code, update readme
