+++ lib/query.js
@@ -3,8 +3,7 @@
  */
 
 var _ = require('@sailshq/lodash'),
-    utils = require('./utils'),
-    hop = utils.object.hasOwnProperty;
+    utils = require('./utils');
 
 /**
  * Query Builder for creating parameterized queries for use
@@ -427,26 +426,6 @@ Query.prototype.sort = function(options) {
   }
 
   this._query += sortItems.join(', ');
-};
-
-/**
- * Specify a `group by` condition
- */
-
-Query.prototype.group = function(options) {
-  var self = this;
-
-  this._query += ' GROUP BY ';
-
-  // Normalize to array
-  if(!Array.isArray(options)) options = [options];
-
-  options.forEach(function(key) {
-    self._query += key + ', ';
-  });
-
-  // Remove trailing comma
-  this._query = this._query.slice(0, -2);
 };
 
 /**
