+++ lib/utils.js
@@ -1,4 +1,4 @@
-const _ = require("underscore");
+const _ = require('@sailshq/lodash');
 
 const utils = module.exports = {};
 
@@ -99,23 +99,6 @@ utils.buildForeignKeyMap = physicalModelsReport => {
 }
 
 /**
-* Safe hasOwnProperty
-*/
-utils.object = {};
-/**
-* Safer helper for hasOwnProperty checks
-*
-* @param {Object} obj
-* @param {String} prop
-* @return {Boolean}
-* @api public
-*/
-var hop = Object.prototype.hasOwnProperty;
-	utils.object.hasOwnProperty = function(obj, prop) {
-	return hop.call(obj, prop);
-};
-
-/**
 * Escape Name
 *
 * Wraps a name in quotes to allow reserved
@@ -136,90 +119,6 @@ utils.buildIndexes = function(obj) {
   });
 
   return indexes;
-};
-
-
-/**
- * Group Results into an Array
- *
- * Groups values returned from an association query into a single result.
- * For each collection association the object returned should have an array
- * under the user defined key with the joined results.
- *
- * @param {Array} results returned from a query
- * @return {Object} a single values object
- */
-utils.group = function(values) {
-
-  var self = this;
-  var joinKeys = [];
-  var _value;
-
-  if (!Array.isArray(values)) return values;
-
-  // Grab all the keys needed to be grouped
-  var associationKeys = [];
-
-  values.forEach(function(value) {
-    Object.keys(value).forEach(function(key) {
-      key = key.split('__');
-      if (key.length === 2) associationKeys.push(key[0].toLowerCase());
-    });
-  });
-
-  associationKeys = _.unique(associationKeys);
-
-  // Store the values to be grouped by id
-  var groupings = {};
-
-  values.forEach(function(value) {
-
-    // Add to groupings
-    if (!groupings[value.id]) groupings[value.id] = {};
-
-    associationKeys.forEach(function(key) {
-      if(!Array.isArray(groupings[value.id][key]))
-        groupings[value.id][key] = [];
-      var props = {};
-
-      Object.keys(value).forEach(function(prop) {
-        var attr = prop.split('__');
-        if (attr.length === 2 && attr[0] === key) {
-          props[attr[1]] = value[prop];
-          delete value[prop];
-        }
-      });
-
-      // Don't add empty records that come from a left join
-      var empty = true;
-
-      Object.keys(props).forEach(function(prop) {
-        if (props[prop] !== null) empty = false;
-      });
-
-      if (!empty) groupings[value.id][key].push(props);
-    });
-  });
-
-  var _values = [];
-
-  values.forEach(function(value) {
-    var unique = true;
-
-    _values.forEach(function(_value) {
-      if (_value.id === value.id) unique = false;
-    });
-
-    if (!unique) return;
-
-    Object.keys(groupings[value.id]).forEach(function(key) {
-      value[key] = _.uniq(groupings[value.id][key], 'id');
-    });
-
-    _values.push(value);
-  });
-
-  return _values;
 };
 
 /**
@@ -393,14 +292,4 @@ utils.sqlTypeCast = function(type, columnName) {
       console.error("Warning: Unregistered type given: " + type);
       return 'TEXT';
   }
-};
-
-/**
- * JS Date to UTC Timestamp
- *
- * Dates should be stored in Postgres with UTC timestamps
- * and then converted to local time on the client.
- */
-utils.toSqlDate = function(date) {
-  return date.toUTCString();
 };
