--- test/unit/query.limit.js
@@ -1,29 +1,0 @@
-var Query = require('../../lib/query'),
-    should = require('should');
-
-describe('query', function() {
-
-  /**
-   * LIMIT
-   *
-   * Adds a LIMIT parameter to a sql statement
-   */
-
-  describe('.limit()', function() {
-
-    // Lookup criteria
-    var criteria = {
-      where: {
-        name: 'foo'
-      },
-      limit: 1
-    };
-
-    it('should append the LIMIT clause to the query', function() {
-      var query = new Query({ name: { type: 'text' }}).find('test', criteria);
-
-      query.query.should.eql('SELECT rowid, * FROM test WHERE LOWER("name") = $1 LIMIT 1');
-    });
-
-  });
-});
