--- test/unit/query.sort.js
@@ -1,69 +1,0 @@
-var Query = require('../../lib/query'),
-    should = require('should');
-
-describe('query', function() {
-
-  /**
-   * SORT
-   *
-   * Adds an ORDER BY parameter to a sql statement
-   */
-
-  describe('.sort()', function() {
-
-    it('should append the ORDER BY clause to the query', function() {
-
-      // Lookup criteria
-      var criteria = {
-        where: {
-          name: 'foo'
-        },
-        sort: {
-          name: 1
-        }
-      };
-
-      var query = new Query({ name: { type: 'text' }}).find('test', criteria);
-      query.query.should.eql('SELECT rowid, * FROM test WHERE LOWER("name") = $1 ORDER BY "name" ASC');
-
-    });
-
-    it('should sort by multiple columns', function() {
-
-      // Lookup criteria
-      var criteria = {
-        where: {
-          name: 'foo'
-        },
-        sort: {
-          name: 1,
-          age: 1
-        }
-      };
-
-      var query = new Query({ name: { type: 'text' }}).find('test', criteria);
-      query.query.should.eql('SELECT rowid, * FROM test WHERE LOWER("name") = $1 ORDER BY "name" ASC, "age" ASC');
-
-    });
-
-    it('should allow desc and asc ordering', function() {
-
-      // Lookup criteria
-      var criteria = {
-        where: {
-          name: 'foo'
-        },
-        sort: {
-          name: 1,
-          age: -1
-        }
-      };
-
-      var query = new Query({ name: { type: 'text' }}).find('test', criteria);
-      query.query.should.eql('SELECT rowid, * FROM test WHERE LOWER("name") = $1 ORDER BY "name" ASC, "age" DESC');
-
-    });
-
-  });
-
-});
